feat(seo): add comprehensive SEO infrastructure
- Add centralized site config (site-config.ts)
- Add robots.txt and sitemap.xml generation
- Add JSON-LD schemas (Person, WebSite, Article)
- Enhance root layout with metadata, schemas, skip-to-content
- Refactor pages to extract client components for server metadata
- Add canonical URLs and OpenGraph metadata to all pages
- Add default OG image
