+++ contentlayer.config.ts
@@ -70,6 +70,7 @@ export const Page = defineDocumentType(() => ({
 
 export default makeSource({
     contentDirPath: "content",
+    disableImportAliasWarning: true,
     documentTypes: [BlogPost, Project, Page],
     mdx: {
         rehypePlugins: [
