+++ eslint.config.mjs
@@ -9,10 +9,14 @@ const eslintConfig = defineConfig([
   globalIgnores([
     // Default ignores of eslint-config-next:
     ".next/**",
+    "**/.next/**",
     "out/**",
     "build/**",
     "next-env.d.ts",
     ".contentlayer/**",
+    "**/.contentlayer/**",
+    ".worktrees/**",
+    "output/**",
   ]),
 ]);
