+++ src/app/layout.tsx
@@ -1,8 +1,11 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono, Instrument_Serif, Outfit } from "next/font/google";
-import "./globals.css";
-import { Navbar } from "@/components/layout/Navbar";
-import { Footer } from "@/components/layout/Footer";
+import type { Metadata } from "next"
+import { Geist, Geist_Mono, Instrument_Serif, Outfit } from "next/font/google"
+import "./globals.css"
+import { Navbar } from "@/components/layout/Navbar"
+import { Footer } from "@/components/layout/Footer"
+import { siteConfig } from "@/lib/site-config"
+import { PersonSchema } from "@/components/seo/PersonSchema"
+import { WebSiteSchema } from "@/components/seo/WebSiteSchema"
 
 
 const geistSans = Geist({
@@ -27,9 +30,35 @@ const outfit = Outfit({
 });
 
 export const metadata: Metadata = {
-  title: "David's Digital Garden",
-  description: "Personal portfolio website of a technology fan",
-};
+  metadataBase: new URL(siteConfig.siteUrl),
+  title: {
+    default: siteConfig.siteName,
+    template: `%s | ${siteConfig.siteName}`,
+  },
+  description: siteConfig.siteDescription,
+  openGraph: {
+    type: "website",
+    locale: "en_AU",
+    url: siteConfig.siteUrl,
+    siteName: siteConfig.siteName,
+    images: [
+      {
+        url: siteConfig.defaultOgImage,
+        width: 1200,
+        height: 630,
+        alt: siteConfig.siteName,
+      },
+    ],
+  },
+  twitter: {
+    card: "summary_large_image",
+    images: [siteConfig.defaultOgImage],
+  },
+  robots: {
+    index: true,
+    follow: true,
+  },
+}
 
 export default function RootLayout({
   children,
@@ -42,8 +71,18 @@ export default function RootLayout({
         className={`${geistSans.variable} ${geistMono.variable} ${instrumentSerif.variable} ${outfit.variable} antialiased min-h-screen flex flex-col bg-background text-foreground`}
         suppressHydrationWarning
       >
+        <PersonSchema />
+        <WebSiteSchema />
+        <a
+          href="#main-content"
+          className="sr-only focus:not-sr-only focus:absolute focus:top-4 focus:left-4 focus:z-50 focus:rounded-md focus:bg-background focus:px-4 focus:py-2 focus:text-sm focus:font-semibold focus:text-foreground focus:shadow-lg"
+        >
+          Skip to content
+        </a>
         <Navbar />
-        <main className="flex-1">{children}</main>
+        <main id="main-content" className="flex-1">
+          {children}
+        </main>
         <Footer />
       </body>
     </html>
