+++ src/app/robots.ts
@@ -1,0 +1,13 @@
+import type { MetadataRoute } from "next"
+import { siteConfig } from "@/lib/site-config"
+
+export default function robots(): MetadataRoute.Robots {
+    return {
+        rules: {
+            userAgent: "*",
+            allow: "/",
+            disallow: ["/api/"],
+        },
+        sitemap: `${siteConfig.siteUrl}/sitemap.xml`,
+    }
+}
