+++ tsconfig.json
@@ -42,6 +42,8 @@
     ".contentlayer/generated"
   ],
   "exclude": [
-    "node_modules"
+    "node_modules",
+    ".worktrees",
+    "output"
   ]
 }
